python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
cryptography==42.0.7
    # via pyopenssl
google-api-core==2.19.0
    # via google-api-python-client
google-api-python-client==2.129.0
    # via pydrive2
google-auth==2.29.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.2.0
    # via google-api-python-client
googleapis-common-protos==1.63.0
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
idna==3.7
    # via requests
oauth2client==4.1.3
    # via pydrive2
proto-plus==1.23.0
    # via google-api-core
protobuf==4.25.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.0
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via
    #   google-auth
    #   oauth2client
pycparser==2.22
    # via cffi
pydrive2==1.19.0
    # via -r requirements.in
pydub==0.25.1
    # via -r requirements.in
pyopenssl==24.1.0
    # via pydrive2
pyparsing==3.1.2
    # via httplib2
python-editor==1.0.4
    # via -r requirements.in
pyyaml==6.0.1
    # via pydrive2
requests==2.31.0
    # via google-api-core
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
send2trash==1.8.3
    # via -r requirements.in
simple-term-menu==1.6.4
    # via -r requirements.in
six==1.16.0
    # via oauth2client
tqdm==4.66.4
    # via -r requirements.in
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.1
    # via requests
